// frontend/src/app/core/api/model/schedule-info-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScheduleInfoResponse { 
    id?: number;
    label?: string;
    status?: ScheduleInfoResponse.StatusEnum;
    promotionId?: number;
}
export namespace ScheduleInfoResponse {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}