// frontend/src/app/core/api/model/time-slot-definition-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeSlotDefinitionResponse { 
    id?: number;
    code?: string;
    label?: string;
    startTime?: string;
    endTime?: string;
    orderIndex?: number;
}